Update Argonsimulation.cpp
Hey,

Josko from Delft here. I was looking at your code and I noted the vel_update() function using v(t+h) = v(t) + h * (F(t) - F(t-h)) / 2,
while formula 8.9b of the ICCP book by J. M. Thijssen.

As your latest commmit said there was a bug in it, I propose this might be it.

Furthermore, I am wondering at a point of style; the files as displayed on github don't contain indentation. Is this a github problem or are you not using it? If you aren't, I would strongly advise you to do so, as it greatly increases the code readability for both yourself and others. Similarly, there is some inconsistency in your variable naming; #341 contains a "Force" while most variables are lowercase, as are most function definitions. This is just nitpicking, but consistency is a good habit.

Cheers,
I hope this helps.

PS: I also hope this is the correct way to suggest code edits :)